Refactor the caret posiiton.
